meta-ampere: mtjade: fix fan react too slow to failure
When a fan is unplugged, tach driver of AST2500 shows a timeout error
in hwmon. TachSensor will treat this as an error and increase the polling
time from 0.5s to 5s to ensure the error. dbus-sensors service needs 5
cycle of error to really treat the sensor as non-functional and update "nan"
to the Value property. Totally, it needs over 25s since when the fan is
unplugged for the sensor to show "nan". Therefore, phosphor-fan can only
react to the failure after that amount of time.
This patch omits changeing the polling time when the error code returned
from reading hwmon is timeout. This helps phosphor-fan react faster to
the fan failure.

Test:

BMC is ON, host is ON
Unplug a fan -> all the other fans will be set to the maximum speed
after approximately 2s.

Signed-off-by: Chau Ly <chaul@amperecomputing.com>