meta-ampere: systemd: prevent writing to RTC
Ampere BMC requirement specifies that BMC should not be able to write to
RTC but just read from it.
This commit adds a patch to prevent IPMI and Redfish commands to write
to the RTC.
Note that Mt.Jade does not have this requirement so add an option to
configure for Mt.Jade to bypass this behavior.

Tested: on Mt.Mitchell:
    1. Read RTC clock at HOST side by command "hwclock -r"
    2. ipmitool sel time set "12/24/2022 12:00:00"
    3. Read RTC clock at HOST side by command "hwclock -r"
    4. Expect the RTC time not change at step#1 and step#3

Signed-off-by: Chanh Nguyen <chanh@os.amperecomputing.com>