meta-ampere: mtmitchell: Support always retrieve BERT when power cycle/reset/off
When BERT crash happens, depend on the Host UEFI firmware configuration,
the Host might not be rebooted immediately so no REBOOT_ACK from Host
for the BMC to check BERT crash. This might cause BERT crash is not
handled.
This commit changes to check BERT every time the Host is powered cycle,
reset or off to handle above case.

Signed-off-by: Dung Cao <dung@os.amperecomputing.com>