meta-ampere: state: Prevent services enter failed state
Scenario:
Sometimes, services can fail to start application in the first time. The
BMC detected that this service entered to failed state, therefore, it
updated the BMC's state to Quiesced. But this service is auto restarted,
and it can successfully after that.

Update:
This commit updated the .service files to prevent services enter the
failed state until it touchs the restart limit conditions.

Tested:
    1. Force services is failed to start in the first time.
    2. Check BMC's state when all of application are ready.
        $busctl get-property xyz.openbmc_project.State.BMC \
        /xyz/openbmc_project/state/bmc0 \
        xyz.openbmc_project.State.BMC CurrentBMCState
    3. The BMC's state are "Ready"

Signed-off-by: Thang Tran <thuutran@amperecomputing.com>