meta-ampere: mctp_i2c_binding: change waiting S1 ready time to 180s
S1 MCTP might take much time to be ready after the S0 MCTP is in ready
state. So, 20s waiting time might not be enough.
This commit changes the time waiting for S1 ready from 20s to 180s. If
the S1 MCTP is still not in ready state, we can assume something wrong
with it.

Tested:
1. Do regression test and stress test with 2P system.
2. No `S1 disabled` issue.

Signed-off-by: Thu Nguyen <thu@os.amperecomputing.com>